Move attrs_of() to core/property_mixins
