+++ bokehjs/src/lib/core/property_mixins.ts
@@ -3,6 +3,7 @@ import {Color} from "./types"
 import {LineJoin, LineCap, FontStyle, HatchPatternType, TextAlign, TextBaseline} from "./enums"
 import * as k from "./kinds"
 import {Texture} from "models/textures/texture"
+import {keys} from "./util/object"
 
 export type HatchPattern = HatchPatternType | string
 export type HatchExtra = {[key: string]: Texture}
@@ -245,3 +246,16 @@ export type OutlineLine = Prefixed<"outline", Line>
 export type SeparatorLine = Prefixed<"separator", Line>
 export type SubGroupText = Prefixed<"subgroup", Text>
 export type TitleText = Prefixed<"title", Text>
+
+type Mixins = Text | Line | Fill | Hatch
+
+export function attrs_of<P extends string, T extends Mixins>(
+    model: any, prefix: P, mixin: p.DefineOf<T>, prefixed: boolean = false): {[key: string]: any} {
+  const attrs: {[key: string]: unknown} = {}
+  for (const attr of keys(mixin)) {
+    const prefixed_attr = `${prefix}${attr}` as const
+    const value = model[prefixed_attr]
+    attrs[prefixed ? prefixed_attr : attr] = value
+  }
+  return attrs
+}
